Si
si

+++ TPV/P2/HolaSDL/Alien.h
@@ -24,7 +24,7 @@ private:
 
 public:
 	Alien();
-	Alien(const Game* gam, Point2D<int> pos, int sub, const Texture* tex, Mothership* mot);
+	Alien(const Game* gam, Point2D<double> pos, int sub, const Texture* tex, Mothership* mot);
 	~Alien();
 	bool update() override;
 	const void render() override;
